+++ src/api/loadConfiguration.ts
@@ -1,0 +1,5 @@
+import { IUserConfiguration } from '../configuration'
+
+export async function loadConfiguration(): Promise<IUserConfiguration> {
+  return {} as IUserConfiguration
+}
