+++ src/models/pickle_order.ts
@@ -1,0 +1,1 @@
+export type PickleOrder = 'defined' | 'random'
